requirements.txt
Flask==2.2.2
tensorflow-cpu==2.8.0
gunicorn==20.1.0
pandas==1.4.2
numpy==1.22.1
pickle-mixin==1.0.2